@@ -1,69 +1,28 @@
 {
-  "name": "mdx-contentlayer-blog",
-  "version": "0.1.0",
+  "name": "joebadcom",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.2.1",
-    "@fortawesome/free-brands-svg-icons": "^6.2.1",
-    "@fortawesome/free-solid-svg-icons": "^6.2.1",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@next/font": "^13.1.1",
-    "@popperjs/core": "^2.11.6",
-    "@tailwindcss/line-clamp": "^0.4.4",
-    "@tailwindcss/typography": "^0.5.3",
-    "animate.css": "^4.1.1",
-    "aws-amplify": "^5.0.10",
-    "contentlayer": "^0.2.5",
-    "date-fns": "^2.29.3",
-    "flowbite": "^1.6.2",
-    "flowbite-react": "^0.3.7",
-    "next": "13.1.1",
-    "next-contentlayer": "^0.2.5",
-    "next-intl": "^2.19.0",
-    "next-seo": "^5.4.0",
-    "prismjs": "^1.29.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-popper": "^2.3.0",
-    "reading-time": "^1.5.0",
-    "rehype-prism-plus": "^2.0.0",
-    "remeda": "^1.24.0",
-    "sass": "^1.63.6",
-    "tw-elements": "^1.0.0-alpha13"
+    "build": "tsx src/build.ts",
+    "dev": "nodemon --watch src --ext md,tsx,jsx,ts,js,css --exec yarn build"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.5",
-    "@netlify/plugin-nextjs": "^4.30.0",
-    "@storybook/addon-actions": "^6.5.14",
-    "@storybook/addon-essentials": "^6.5.14",
-    "@storybook/addon-interactions": "^6.5.14",
-    "@storybook/addon-links": "^6.5.14",
-    "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/builder-webpack5": "^6.5.14",
-    "@storybook/manager-webpack5": "^6.5.14",
-    "@storybook/react": "^6.5.14",
-    "@storybook/testing-library": "^0.0.13",
-    "@types/node": "^18.11.9",
-    "@types/react": "^18.0.25",
-    "autoprefixer": "^10.4.7",
-    "babel-loader": "^8.3.0",
-    "eslint": "8.19.0",
-    "eslint-config-next": "12.2.0",
-    "eslint-plugin-storybook": "^0.6.8",
-    "postcss": "^8.4.14",
-    "tailwindcss": "^3.1.5",
-    "typescript": "^4.9.3",
-    "util": "^0.12.5"
-  },
-  "resolutions": {
-    "webpack": "^5"
+    "@tailwindcss/cli": "^4.1.10",
+    "@tailwindcss/typography": "^0.5.16",
+    "@types/node": "^24.0.4",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "fast-xml-parser": "^5.2.5",
+    "glob": "^11.0.3",
+    "gray-matter": "^4.0.3",
+    "nodemon": "^3.1.10",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "remark": "^15.0.1",
+    "remark-html": "^16.0.1",
+    "tailwindcss": "^4.1.10",
+    "tsx": "^4.20.3",
+    "typescript": "~5.8.3"
   }
 }
